refactor & add more e2e tests for array_max
