+++ src/expr/src/vector_op/array_min_max.rs
@@ -18,6 +18,8 @@ use risingwave_expr_macro::function;
 
 use crate::Result;
 
+/// FIXME: #[`function("array_min(list`) -> any")] supports
+/// In this way we could avoid manual macro expansion
 #[function("array_min(list) -> *int")]
 #[function("array_min(list) -> *float")]
 #[function("array_min(list) -> decimal")]
@@ -31,6 +33,25 @@ use crate::Result;
 #[function("array_min(list) -> bytea")]
 pub fn array_min<T: Scalar>(list: ListRef<'_>) -> Result<Option<T>> {
     let min_value = list.iter().flatten().map(DefaultOrdered).min();
+    match min_value.map(|v| v.0).to_owned_datum() {
+        Some(s) => Ok(Some(s.try_into()?)),
+        None => Ok(None),
+    }
+}
+
+#[function("array_max(list) -> *int")]
+#[function("array_max(list) -> *float")]
+#[function("array_max(list) -> decimal")]
+#[function("array_max(list) -> serial")]
+#[function("array_max(list) -> int256")]
+#[function("array_max(list) -> date")]
+#[function("array_max(list) -> time")]
+#[function("array_max(list) -> timestamp")]
+#[function("array_max(list) -> timestamptz")]
+#[function("array_max(list) -> varchar")]
+#[function("array_max(list) -> bytea")]
+pub fn array_max<T: Scalar>(list: ListRef<'_>) -> Result<Option<T>> {
+    let min_value = list.iter().flatten().map(DefaultOrdered).max();
     match min_value.map(|v| v.0).to_owned_datum() {
         Some(s) => Ok(Some(s.try_into()?)),
         None => Ok(None),
