+++ src/expr/src/vector_op/mod.rs
@@ -13,11 +13,10 @@
 // limitations under the License.
 
 pub mod arithmetic_op;
-pub mod array_max;
 pub mod array_access;
 pub mod array_distinct;
 pub mod array_length;
-pub mod array_min;
+pub mod array_min_max;
 pub mod array_positions;
 pub mod array_range_access;
 pub mod array_remove;
